Implement job / process control according to libc documentation
Follow the libc documentation in accordance with process and job
control.  Although I plan on extending this functionality to support
parallel pipes, this provides the correct matching functionality for
existing shells.

This also adds a dependency on the Omni native PHP extension, which
provides some additional native C functions which are not available
through the POSIX or Process Control extensions.